PDF_Magic/src/core/__init__.py
# Autor: Leon Gajtner
# Datum: 07.11.2024
# Version: 2.1
# Automatisch generierte __init__.py für core

# Importiere Module
from .converter import *
from .file_handler import *
from .utils import *
from .validator import *

# Core-spezifische Importe
from PyPDF2 import PdfReader, PdfWriter
from docx import Document

__all__ = [
    'converter',
    'file_handler',
    'utils',
    'validator',
]
