PDF_Magic/requirements.txt
PyQt5>=5.15.0
PyPDF2>=3.0.0
python-docx>=0.8.11